# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MADHack Agon API",
    "description": "MADHack Initial Round API Documentation",
    "contact": {
      "email": "ucscieee@gmail.com"
    },
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "https://agon.madhack.ucscieee.com/api",
      "description": "Agon Server"
    },
    {
      "url": "https://backup.agon.madhack.ucscieee.com/api",
      "description": "Agon Backup Server"
    },
    {
      "url": "http://localhost:3001/api",
      "description": "Local Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/user": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "retrieves user details",
        "description": "Gets details of logged in user",
        "operationId": "getUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserWithoutPassword"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_400_object"
                },
                "example": {
                  "error": "fieldName is required"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_401_object"
                },
                "example": {
                  "error": "No auth token provided"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "updates user details",
        "description": "Updates details of logged in user",
        "operationId": "updateUser",
        "requestBody": {
          "description": "User details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppUserWithoutPasswordAndId"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_200_object"
                },
                "example": {
                  "message": "User updated successfully"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_400_object"
                },
                "example": {
                  "error": "fieldName is required"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_401_object"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_500_object"
                },
                "examples": {
                  "internal-error": {
                    "value": {
                      "error": "Internal Server Error"
                    }
                  },
                  "update-error": {
                    "value": {
                      "error": "Error updating user"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "deletes user",
        "description": "Deletes logged in user",
        "operationId": "deleteUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_200_object"
                },
                "example": {
                  "message": "User deleted successfully"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_401_object"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_500_object"
                },
                "examples": {
                  "internal-error": {
                    "value": {
                      "error": "Internal Server Error"
                    }
                  },
                  "delete-error": {
                    "value": {
                      "error": "Error deleting user"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/user/upload-avatar": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "uploads user avatars",
        "description": "Uploads images to a S3 bucket and returns the path",
        "operationId": "uploadAvatar",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/user_upload_avatar_req_body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user_upload_avatar_res_body"
                },
                "example": {
                  "message": "Image uploaded successfully",
                  "avatarUrl": "/user-avatars/064596AB-85CE-4107-B390-2E3989F7EF40.png"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_400_object"
                },
                "example": {
                  "error": "No image selected"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_401_object"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_500_object"
                },
                "examples": {
                  "internal-error": {
                    "value": {
                      "error": "Internal Server Error"
                    }
                  },
                  "upload-error": {
                    "value": {
                      "error": "Error uploading avatar"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppUserWithoutPassword": {
        "required": [
          "userId",
          "firstName",
          "lastName",
          "email",
          "contactNo",
          "avatarUrl"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "johndoe@example.com"
          },
          "contactNo": {
            "type": "string",
            "example": "0712345678"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "example": "user-avatars/d290f1ee-6c54-4b01-90e6-d701748f0851.png"
          }
        },
        "description": "A user of the mobile application"
      },
      "AppUserWithoutPasswordAndId": {
        "required": [
          "firstName",
          "lastName",
          "email",
          "contactNo",
          "avatarUrl"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "johndoe@example.com"
          },
          "contactNo": {
            "type": "string",
            "example": "0712345678"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "example": "user-avatars/d290f1ee-6c54-4b01-90e6-d701748f0851.png"
          }
        },
        "description": "A user of the mobile application"
      },
...
      "response_200_object": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "response_401_object": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "API access denied! No team access token provided"
          }
        }
      },
      "response_400_object": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        }
      },
...
      "response_500_object": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Internal Server Error"
          }
        }
      },
...
      "user_upload_avatar_req_body": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          }
        }
      },
      "user_upload_avatar_res_body": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "X-API-Key",
        "in": "header"
      }
    }
  }
}